🚚 Move : Graph.js, Search.js 이동 및 적용
- Components안에 폴더로 나뉘어져있던 것이 폴더별 하나의 파일만 있었기에 Components에서 관리하도록 수정
