+++ src/Pages/RecentPrice/Components/Search.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components'
-import { WidthAutoCSS } from '../../../../Styles/common'
+import { WidthAutoCSS } from '../../../Styles/common'
 import { FaSearch } from 'react-icons/fa'
-import useInput from '../../../../Hooks/useInput'
+import useInput from '../../../Hooks/useInput'
 
 function RecentSearch({ onSearchPrd }) {
 	const [search, onChange] = useInput('')
