+++ src/Pages/RecentPrice/index.js
@@ -2,9 +2,8 @@ import { useState } from 'react'
 import styled from 'styled-components'
 import { WidthAutoCSS } from '../../Styles/common'
 import recentPriceMock from '../../__mock__/Data/Product/recentPrice.data'
-
-import Graph from './Components/Graph/Graph'
-import RecentSearch from './Components/Search/Search'
+import Graph from './Components/Graph'
+import RecentSearch from './Components/Search'
 
 function RecentPrice() {
 	const [dummyData, setDummyData] = useState(recentPriceMock)
